[Common] New class ElementMapper
ElementMapper is an optimized variant of a Mapper
that always indexes all codim 0 entities and uses a
GeometryTypeIndex-based vector for GeometryType
index offset lookup instead of a map.

This class is mainly a utility for the global assembler
and for classes that cache information per grid cell
(e.g. the new VTK output infrastructure).

Imported from SVN: r1938 at branches/new_dof_handling