[TypeTree] Change algorithm for variadic template child extraction in transformation engine
This patch switches the child extraction algorithm for the VariadicCompositeNode from
a hand-written recursion over the children to a compiler-generated iteration over the
children. Apart from greatly improving the readibility of the code, this also has the
advantage of not polluting the call stack with lots of intermediate function calls which
do nothing but extract a single child from the node.

Imported from SVN: r1968 at branches/new_dof_handling