[Merge] Forward-port clock functionality (r2050-2053) from trunk
The patches adding the new clock functionality also
introduce a PDELab library. Having a library in trunk,
but not in the branch makes switching between the two
unnecessarily difficult, so this patch forward-ports
the functionality and in particular the addition of the
library.

- [clock] Functions to get wall time and process time and a class to store times
- [hostname] Mode C++-Friendly hostname wrapper here from Dune-PM.
- [logtag] Move to PDELab (from PM).
- [LocalTag] Class for easy boilerplate output inside functions.


Imported from SVN: r2095 at branches/new_dof_handling