[multiindex.hh] Split multiindex.hh into separate files for MultiIndex, DOFIndex and the simplified infrastructure
multiindex.hh has become a *little* big for comfort, and when adding GlobalDOFIndex to
the zoo in there, I decided that enough is enough.

This patch splits the file into one for the MultiIndex, one for the DOFIndex (which
includes the MultiIndex) and a completely separate one for the simplified infrastructure.