[Unordered Containers] Import unordered containers from C++11 or TR1 into PDELab namespace
Since replacing several std::maps with unordered_maps, PDELab has a hard dependency on
C++11 because unordered_map is always imported from std::. For backwards compatibility,
it would be nice to also support the implementation from TR1 (and, possibly, boost).

This patch creates the necessary infrastructure for doing that by testing for both implementations
and importing one into the PDELab namespace, as has become a dear tradition for Dune.