[Utility] Add helper TMP for mapping (dimension,topologyId) -> GeometryType::BasicType
This patch adds a little helper called BasicTypeFromDimensionAndTopologyId to get the
BasicType that corresponds to a given pair of dimension and topologyId.

In theory, it is not hard to do that conversion, but the process is a little awkward and requires
internals of GenericGeometry, construction of the corresponding topology and knowledge of the
BasicType used for lines. The information provided by this helper is mostly useful when writing
generic front ends for FiniteElementMaps. For many grids, those front ends can extract a valid
topologyId from the hasSingleGeometryType capability of the grid, freeing users from having to
specify the BasicType of the grid cells manually.