[TypeTree] Fix compilation error when trying to transform FilteredCompositeNode
When trying to transform a FilteredCompositeNode, the transformation algorithm
reuses the plugins for the VariadicCompositeNode (in fact, the FilteredCompositeNode
even exports the node tag of the VariadicCompositeNode). This causes a problem because
the VariadicCompositeNode code tries to call nodeStorage() to construct a matching index
pack. That method does not exist for the FilteredCompositeNode, so this patch adds
some gymnastics to build the index pack in a way that also works with the filtered
node.