[Cleanup] Remove legacy code for older compilers
This commit removes a major amount of code that was required to support
compilers without support for variadic templates, rvalue references or
SFINAE on constructors. In particular, with this commit the minimum
version requirement of dune-typetree is bumped to GCC 4.5 or compatible.

As part of this change, VariadicCompositeNode and its accompanying tag
were renamed to CompositeNode(Tag), so users *will* have to update their code!