[bugfix,pkg-config] Add library to pkg-config file.
Oviously this module ships a library but lacks it in the linker flags
for pkg-config. Therefore the build of modules that depend on an
installed version of typetree fail miserably with unresolved symbols
when the library is checked.

This patch solves this by adding the -ldunetypetree to the linker flags.
It should be applied to the release, of course. Otherwise testing _will_
fail.