[Utility] Add namespace with predefined index_constants
If the exact value of a compile time index is known to the programmer,
it is convenient not to have to construct an index_constant for that
value by hand (cf. std::bind and its placeholders).

This patch adds predefined index_constants for the range [0,19] in a new
namespace Dune::TypeTree::Indices. Those constants are constexpr, which
has the benefits of making them usable in compile time expressions (e.g.
template parameters) and avoiding ODR problems due to the internal
linkage of constexpr objects.