[child extraction] Allow alternative syntax for static child access
This patch modifies the compile time accessors for direct children to
accept an index_constant as function parameter in addition to specifying
the index using traditional angled brackets. This keeps the generic
child extraction mechanism much cleaner and gives users a chance to
avoid the pesky angled brackets, which typically require them to
explicitly write 'template' in front of the function name, which is
especially nice in combination with the predefined constants like _1
etc.

With this patch, the following two calls are equivalent:

node.template child<i>();
node.child(index_constant<i>{});