[child extraction] Provide new overloads for child() methods in nodes
Calling the freestanding child() function in user code is often
inconvenient because the call has to be qualified due to the fact that
the actual node type is typically a derived type and does not live in
the TypeTree namespace.

This patch adds new overloads for the child() methods on PowerNode and
CompositeNode, which forward their arguments to the freestanding
function. With this patch, the two following lines are equivalent:

Dune::TypeTree::child(my_node,_0,3,_2);
my_node.child(_0,3,_2);