[child extraction] Complain when member function child() is used incorrectly
The member function child() of PowerNode and CompositeNode can only work
correctly if the TreePath is not empty because it actually has to return
a child of the current node; if it returns the node itself, it gets
sliced to tbe type of the TypeTree base class.

So there are now static assertions that give the user a hint what to do
in this case.