[child extraction]
Fix the static assertion in the child() member functions. I didn't pay
attention to the fact that logical operator short circuiting doesn't
help there, so we must make sure not to call treePathSize on a
std::size_t or an index_constant by accident.

This patch adds a little helper function that implements that switch
using SFINAE and updates the child() functions to use that bridge.

Thanks to Oliver for reporting!