Calm done clang by proper constexpr values.
In contratst to gcc, clang does not consider the result
of a static constexpr function called for a non-constexpr
instance of the class a constexpr. I don't know who's right,
but we can simply avoid this by calling the methods on the
class directly. Unfortunately this requires a little more
boiler-plate code to get rid of references.