Deprecate TreePath* helper classes
These classes implement common operations like push_back...
for purely static `HybridTreePath`s. We decided that they should
be removed, since they duplicate the functionality of the free
functions.

Since we would get a deprecation warning if we specialize
a deprecated class template, this only adds the `[[deprecated("...")]]`
attribute to the specialization and uses doxygen to deprecate
the class template declaration.