Add alias to deduce the direct children
For nodes satisfying `StaticDegreeInnerTreeNode` but not
`UniformInnerTreeNode` the lack of the `Childtypes` member typedef
(that exists in the old base `CompositeNode` base classs) makes the
handling of such nodes cumbersome in user code.

Since we don't want to introduce the typedef in the concepts
to keep them as slim as needed, this introduces a helper typedef
`Dune::TypeTree::Impl::Children<Node>` that provides a tuple of child
types deduced using `Dune::TypeTree::Child<Node, i>`.

For nodes where this is not possible this is an alias for `void`.
One could alternativly use an empty `std::tuple<>`, but this
would give the false impression, that there are no children.