[test] Add a rudimentary test for Child extraction
This adds a rudimentary test for `Child`, `ChildForTreePath`,
`Impl::Children`, and `child`.

The test builds a tree for testing and checks by comparing
these utilties for all nodes against explicitly given correct
results.

While one could write a generic test, that traverses the tree
generically, this would come with the danger, that the test
code is incorrect. E.g. a check that deduces the correct child
tuple from `Child` to check correctness of `Children` would
essentially duplicate the implementation of `Children`.