Move child,Child,ChildForTreePath,Children to common/typetree/childacces.hh
This introduces a new directory `dune/common/typetree/` and extracts
the modern child access utilities to the new file `childaccess.hh`
within this folder.

The short term plan for 2.11 is to make `dune/common/typetree/`
self-contained and independent of the rest of dune-typetree.
The mid-term plan is t eventually move the content of this
directory to dune-common. Ideally this would not even require
modifications to user code that only uses the modernized part
of typetree from `dune/common/typetree/`.