Move hybridmultiindex.hh to dune/common/...
Since this was not part of any release yet, there's no need
for a fallback header. However, at least dune-pdelab master
uses it in one place, so we should not break this code.
Maybe we can remove the deprecated dummy
`dune/typetree/hybridmultiindex.hh` before the next release.

This deliberately goes into `dune/common/` and not
`dune/common/typetree/` as decided during because the
code lives in `Dune::`.