[cleanup] Remove dune/typetree/hybridmultiindex.hh
This is a follow-up to !152 which has moved the header to
`dune/common/hybridmultiindex.hh`. The deprecated old header
`dune/typetree/hybridmultiindex.hh` was only kept to not
break tests in downstream moduls during the transition.
Once downstream has adjusted to the new location the present
MR can be merged to remove the header _before_ the next release.