Introduce TreePath as alias for HybridMultiIndex
There is the long existing plan that the name `TreePath`
should eventually replace `HybridTreePath`. This was originally
not possible, because there existed another class `TreePath`
before. Following out transition plan (that we almost forgot about)
this is long gone and we're free to reuse the name now.

We should do this before moving to dune-common and not introduce
the old `HybridTreePath` there.